fix: 修复macOS arm64编译失败
